services/train/requirements.txt
psycopg2-binary
mlflow
pandas
python-dotenv
scikit-learn